+++ bin/csv-filter
@@ -14,13 +14,15 @@ Usage: #{parser.program_name} [options]
   Reads and parses the CSV text content of the standard input per the given input options.
   From that content, generates CSV text per the given output options
   and writes that text to the standard output.
-
 BANNER
 
-parser.separator('Generic Options')
-parser.separator(nil)
-
-parser.parse!
+begin
+  parser.parse!
+rescue OptionParser::InvalidOption
+  $stderr.puts($!.message)
+  $stderr.puts(parser)
+  exit(false)
+end
 
 CSV.filter(**options) do |row|
 end
